flex-algo: extract protocol-neutral core out of isis/
First step toward OSPFv2/v3 Flexible Algorithm (RFC 9350) support: move
the parts of the IS-IS flex-algo implementation that are not
IS-IS-specific into shared homes so OSPF can reuse them, with no
behavior change.

- packet-utils: host ExtAdminGroup (RFC 7308 admin-group bitmap) with
  reusable parse_be/emit; isis-packet re-exports it so
  isis_packet::ExtAdminGroup keeps resolving. Both isis-packet and the
  forthcoming OSPF FAD codec share one bitmap type.
- zebra-rs: new crate::flex_algo module holding the protocol-neutral
  FlexAlgoEntry / FadMetricType data model and the FAD link-admission
  constraint engine (link_passes_fad + admin-group set ops). Name->bit
  resolution is abstracted behind a small AffinityBits trait so the
  engine doesn't depend on any protocol's affinity table.
- isis: AffinityMap implements AffinityBits; isis::flex_algo re-exports
  the moved types so graph.rs / rib.rs call sites are unchanged.

The IS-IS config-staging machinery (FlexAlgoConfig, AffinityMap, SRLG)
and the isis-packet FAD/ASLA builders stay put; the affinity-map/SRLG
tables move to global scope in a follow-up PR (they change IS-IS config
paths). Pure refactor: workspace clippy + fmt clean, all flex-algo /
affinity / packet tests pass.
