+++ crates/packet-utils/src/lib.rs
@@ -1,9 +1,11 @@
+mod admin_group;
 mod algo;
 mod many0;
 mod safe_split_at;
 mod sid_label;
 mod util;
 
+pub use admin_group::*;
 pub use algo::*;
 pub use many0::*;
 pub use safe_split_at::*;
