+++ zebra-rs/src/isis/flex_algo.rs
@@ -1,5 +1,4 @@
 use std::collections::{BTreeMap, BTreeSet};
-use std::str::FromStr;
 
 use anyhow::{Context, Result, bail};
 use isis_packet::neigh::IsisSubTlv as NeighSubTlv;
@@ -15,56 +14,10 @@ use super::Isis;
 use super::affinity_map::AffinityMap;
 use super::srlg::SrlgGroup;
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum FadMetricType {
-    Igp,                // FAD Metric-Type 0
-    MinUnidirLinkDelay, // FAD Metric-Type 1 (RFC 8570)
-    TeDefault,          // FAD Metric-Type 2 (RFC 5305)
-}
-
-impl FadMetricType {
-    /// FAD Sub-TLV Metric-Type code (RFC 9350 §5.1, IANA registry).
-    /// Single source of truth for the on-the-wire byte, consumed by
-    /// `build_fad_subs` at LSP-build time.
-    pub fn wire(self) -> u8 {
-        match self {
-            Self::Igp => 0,
-            Self::MinUnidirLinkDelay => 1,
-            Self::TeDefault => 2,
-        }
-    }
-}
-
-impl FromStr for FadMetricType {
-    type Err = anyhow::Error;
-    fn from_str(s: &str) -> Result<Self> {
-        match s {
-            "igp" => Ok(Self::Igp),
-            "min-unidir-link-delay" => Ok(Self::MinUnidirLinkDelay),
-            "te-default" => Ok(Self::TeDefault),
-            _ => bail!("unknown flex-algo metric-type: {s}"),
-        }
-    }
-}
-
-/// One Flexible Algorithm Definition (RFC 9350) as configured on this
-/// router. Mirrors the YANG schema under /router/isis/flex-algo.
-#[derive(Debug, Default, Clone)]
-pub struct FlexAlgoEntry {
-    pub delete: bool,
-    pub advertise_definition: Option<bool>,
-    pub metric_type: Option<FadMetricType>,
-    pub priority: Option<u8>,
-    pub prefix_metric: Option<bool>,
-    pub dataplane_sr_mpls: Option<bool>,
-    pub dataplane_srv6: Option<bool>,
-    pub dataplane_ip: Option<bool>,
-    pub include_any: BTreeSet<String>,
-    pub include_all: BTreeSet<String>,
-    pub exclude_any: BTreeSet<String>,
-    pub srlg_exclude: BTreeSet<String>,
-    pub ti_lfa: bool,
-}
+// The protocol-neutral flex-algo data model + FAD constraint engine
+// now live in `crate::flex_algo`; re-export so existing IS-IS call
+// sites (`super::flex_algo::…` in graph.rs / rib.rs) keep resolving.
+pub use crate::flex_algo::{FadMetricType, FlexAlgoEntry, link_passes_fad, local_link_affinity};
 
 pub struct FlexAlgoConfig {
     pub config: BTreeMap<u8, FlexAlgoEntry>,
@@ -205,106 +158,6 @@ pub fn build_link_asla(affinity: &BTreeSet<String>, am: &AffinityMap) -> Option<
         udabm: Vec::new(),
         subs: vec![NeighSubTlv::AdminGrp(IsisSubAdminGrp { groups: words })],
     })
-}
-
-/// Resolve a set of affinity names to an `ExtAdminGroup` bitmap via
-/// the affinity-map. Names with no matching entry are silently
-/// dropped (best-effort, matches `build_link_asla` semantics).
-///
-/// Used by per-algo SPF to derive the bitmap for our own ExtIsReach
-/// edges, since peer-ingested `peer_link_affinity` deliberately
-/// excludes the local sys-id (the rebuild skips self).
-pub fn local_link_affinity(affinity: &BTreeSet<String>, am: &AffinityMap) -> ExtAdminGroup {
-    let mut g = ExtAdminGroup::default();
-    for name in affinity {
-        if let Some(bit) = am.bit(name) {
-            g.set(bit);
-        }
-    }
-    g
-}
-
-/// Apply the RFC 9350 §6 link-attribute constraints from `entry`
-/// against `affinity`. Returns true when the link is admissible for
-/// the algorithm's SPF graph.
-///
-/// `affinity = None` means the source LSP did not advertise an ASLA
-/// bitmap for this neighbor — treated as the empty bitmap (every bit
-/// = 0). That's the right default for peers that simply haven't
-/// configured admin-groups: include-any with a non-empty constraint
-/// rejects them, which matches the §6 "no link attribute" reading.
-///
-/// Constraint semantics:
-///   - **exclude-any**: link fails if any of the FAD's excluded bits
-///     is set in `affinity` (intersection non-empty).
-///   - **include-any**: when the FAD lists any bit here, the link
-///     must have at least one of them set (intersection non-empty).
-///     Empty constraint = no requirement.
-///   - **include-all**: every bit in the FAD's constraint must be set
-///     on the link.
-///
-/// Name resolution failures (a constraint name not in `am`) are
-/// silently dropped, matching `build_fad_subs` — the wire form would
-/// not have carried that bit anyway.
-pub fn link_passes_fad(
-    affinity: Option<&ExtAdminGroup>,
-    entry: &FlexAlgoEntry,
-    am: &AffinityMap,
-) -> bool {
-    let exclude = local_link_affinity(&entry.exclude_any, am);
-    let include_any = local_link_affinity(&entry.include_any, am);
-    let include_all = local_link_affinity(&entry.include_all, am);
-
-    let empty = ExtAdminGroup::default();
-    let bitmap = affinity.unwrap_or(&empty);
-
-    if !ext_admin_group_intersection(&exclude, bitmap).is_empty() {
-        return false;
-    }
-    if !include_any.words.iter().all(|w| *w == 0)
-        && ext_admin_group_intersection(&include_any, bitmap).is_empty()
-    {
-        return false;
-    }
-    if !ext_admin_group_contains(bitmap, &include_all) {
-        return false;
-    }
-    true
-}
-
-/// Bitwise AND of two `ExtAdminGroup` bitmaps. Returned bitmap is
-/// length min(a, b) — trailing zero words from a longer operand do
-/// not contribute set bits.
-fn ext_admin_group_intersection(a: &ExtAdminGroup, b: &ExtAdminGroup) -> ExtAdminGroup {
-    let len = a.words.len().min(b.words.len());
-    let mut out = Vec::with_capacity(len);
-    for i in 0..len {
-        out.push(a.words[i] & b.words[i]);
-    }
-    ExtAdminGroup { words: out }
-}
-
-/// True iff every set bit in `needed` is also set in `bitmap`.
-/// Trailing words past `bitmap.words.len()` in `needed` must be zero.
-fn ext_admin_group_contains(bitmap: &ExtAdminGroup, needed: &ExtAdminGroup) -> bool {
-    for (i, w) in needed.words.iter().enumerate() {
-        let have = bitmap.words.get(i).copied().unwrap_or(0);
-        if *w & !have != 0 {
-            return false;
-        }
-    }
-    true
-}
-
-/// True iff `g` has no set bits.
-trait ExtAdminGroupExt {
-    fn is_empty(&self) -> bool;
-}
-
-impl ExtAdminGroupExt for ExtAdminGroup {
-    fn is_empty(&self) -> bool {
-        self.words.iter().all(|w| *w == 0)
-    }
 }
 
 /// Build the per-algorithm Prefix-SID sub-TLVs (RFC 8667 §2.1 +
@@ -1230,127 +1083,5 @@ mod tests {
             .unwrap_err()
             .to_string();
         assert!(err.contains("128..=255"), "unexpected err: {err}");
-    }
-
-    /// Build a synthetic affinity-map with the requested name → bit
-    /// assignments. Each name gets its own slot starting at bit 0.
-    fn affinity_map(names: &[&str]) -> AffinityMap {
-        let mut am = AffinityMap::new();
-        for (i, name) in names.iter().enumerate() {
-            am.config.insert(
-                (*name).to_string(),
-                super::super::affinity_map::AffinityEntry {
-                    delete: false,
-                    bit_position: Some(i as u16),
-                },
-            );
-        }
-        am
-    }
-
-    /// Build an `ExtAdminGroup` from a list of bit positions.
-    fn admin_group(bits: &[u16]) -> ExtAdminGroup {
-        let mut g = ExtAdminGroup::default();
-        for b in bits {
-            g.set(*b);
-        }
-        g
-    }
-
-    fn fad_entry(
-        exclude_any: &[&str],
-        include_any: &[&str],
-        include_all: &[&str],
-    ) -> FlexAlgoEntry {
-        FlexAlgoEntry {
-            exclude_any: affinity_set(exclude_any),
-            include_any: affinity_set(include_any),
-            include_all: affinity_set(include_all),
-            ..Default::default()
-        }
-    }
-
-    #[test]
-    fn link_passes_fad_no_constraints_accepts_anything() {
-        let am = affinity_map(&["red"]);
-        let entry = fad_entry(&[], &[], &[]);
-        assert!(link_passes_fad(None, &entry, &am));
-        let g = admin_group(&[0]);
-        assert!(link_passes_fad(Some(&g), &entry, &am));
-    }
-
-    #[test]
-    fn link_passes_fad_exclude_any_drops_link_with_excluded_bit() {
-        let am = affinity_map(&["red", "blue"]);
-        let entry = fad_entry(&["red"], &[], &[]);
-        let red = admin_group(&[0]);
-        let blue = admin_group(&[1]);
-        assert!(!link_passes_fad(Some(&red), &entry, &am));
-        assert!(link_passes_fad(Some(&blue), &entry, &am));
-        // Missing affinity = empty bitmap = no excluded bit set.
-        assert!(link_passes_fad(None, &entry, &am));
-    }
-
-    #[test]
-    fn link_passes_fad_include_any_requires_at_least_one_bit() {
-        let am = affinity_map(&["red", "blue", "green"]);
-        let entry = fad_entry(&[], &["red", "blue"], &[]);
-        // No bits set → fails include-any when constraint is non-empty.
-        assert!(!link_passes_fad(None, &entry, &am));
-        // Unrelated bit only → still fails.
-        let green = admin_group(&[2]);
-        assert!(!link_passes_fad(Some(&green), &entry, &am));
-        // One of the required bits → passes.
-        let red = admin_group(&[0]);
-        assert!(link_passes_fad(Some(&red), &entry, &am));
-        // Both required bits → passes.
-        let red_blue = admin_group(&[0, 1]);
-        assert!(link_passes_fad(Some(&red_blue), &entry, &am));
-    }
-
-    #[test]
-    fn link_passes_fad_include_all_requires_every_bit() {
-        let am = affinity_map(&["red", "blue", "green"]);
-        let entry = fad_entry(&[], &[], &["red", "blue"]);
-        // Empty bitmap → missing both → fails.
-        assert!(!link_passes_fad(None, &entry, &am));
-        // Only one of the required bits → fails.
-        let red = admin_group(&[0]);
-        assert!(!link_passes_fad(Some(&red), &entry, &am));
-        // Both required bits → passes.
-        let red_blue = admin_group(&[0, 1]);
-        assert!(link_passes_fad(Some(&red_blue), &entry, &am));
-        // Superset (all required + extra) → still passes.
-        let red_blue_green = admin_group(&[0, 1, 2]);
-        assert!(link_passes_fad(Some(&red_blue_green), &entry, &am));
-    }
-
-    #[test]
-    fn link_passes_fad_combined_constraints_all_must_pass() {
-        // exclude red, include-any {blue, green}, include-all {blue}
-        let am = affinity_map(&["red", "blue", "green"]);
-        let entry = fad_entry(&["red"], &["blue", "green"], &["blue"]);
-        // red present → exclude trips first.
-        let red_blue = admin_group(&[0, 1]);
-        assert!(!link_passes_fad(Some(&red_blue), &entry, &am));
-        // blue alone → satisfies include-any (blue ∈ {blue, green}) and
-        // include-all ({blue} ⊆ {blue}).
-        let blue = admin_group(&[1]);
-        assert!(link_passes_fad(Some(&blue), &entry, &am));
-        // green alone → satisfies include-any but not include-all.
-        let green = admin_group(&[2]);
-        assert!(!link_passes_fad(Some(&green), &entry, &am));
-    }
-
-    #[test]
-    fn link_passes_fad_unresolved_constraint_names_silently_drop() {
-        // FAD references "purple" which isn't in the affinity-map; the
-        // bit cannot be encoded into the local bitmap, so an include-
-        // all on it alone reduces to an empty requirement → passes.
-        let am = affinity_map(&["red"]);
-        let entry = fad_entry(&[], &[], &["purple"]);
-        assert!(link_passes_fad(None, &entry, &am));
-        let red = admin_group(&[0]);
-        assert!(link_passes_fad(Some(&red), &entry, &am));
     }
 }
